+++ src/systemEntities/email.js
@@ -3,6 +3,11 @@
  */
 'use strict';
 
+/** @typedef {import('../wingbot/CustomEntityDetectionModel').EntityDetector} EntityDetector */
+/** @typedef {import('../wingbot/CustomEntityDetectionModel').DetectorOptions} DetectorOptions */
+
+/** @type {[string,EntityDetector|RegExp,DetectorOptions]} */
 module.exports = ['email', /(?<=(\s|^|:))[a-zA-Z0-9!#$%&'*+\-=?^_`{|}~"][^@:\s]*@[^.@\s]+\.[^@\s,]+/, {
-    anonymize: true
+    anonymize: true,
+    clearOverlaps: true
 }];