+++ src/systemEntities/phone.js
@@ -3,11 +3,16 @@
  */
 'use strict';
 
+/** @typedef {import('../wingbot/CustomEntityDetectionModel').EntityDetector} EntityDetector */
+/** @typedef {import('../wingbot/CustomEntityDetectionModel').DetectorOptions} DetectorOptions */
+
+/** @type {[string,EntityDetector|RegExp,DetectorOptions]} */
 module.exports = [
     'phone',
     /((00|\+)[\s-]?[0-9]{1,4}[\s-]?)?([0-9]{3,4}[\s-]?([0-9]{2,3}[\s-]?[0-9]{2}[\s-]?[0-9]{2,3}|[0-9]{3,4}[\s-]?[0-9]{3,4}))(?=(\s|$|[,!.?\-:]))/,
     {
         anonymize: true,
+        clearOverlaps: true,
         extractValue: (match) => {
             let [, internat,, number] = match;