+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.43.0] - 2022-11-17
+
+### Added
+
+- `ai.registerEntityDetector()` registers detector to all known models any time
+- `ai.configureEntityDetector()` configures all known models any time
+
+### Fixed
+
+- local entities are now detected with overlaps
+
 ## [3.42.1] - 2022-11-04
 
 ### Fixed